Added Link Script for Self-Development Page
The Self-Development page can now take URL parameters that determine which panels are open when the page loads. These parameters are used by the back links in the Eloquent JavaScript 2024 page, to make the page accessed by the back links less confusing by having the panel with the link to the Eloquent JavaScript open when the Self-Development page loads.