Eloquent JavaScript 2024 - Ch 20 Advanced Project
Completed a micro-project exercise in Node.js and browser JavaScript. As this project runs its own live backend system, its files are not included in the front-end folder that's deployed to AWS. The details of the exercise are within in the exercise itself, under the "nodejs/live-filesystem" folder. Also included .gitignore data for Node.