Integrated Archive & Removed WIP Page
The Archive page is now the homepage link in the place of the WIP page, and WIP page is now the first subpage of the Archive page. Any other archived pages are planned to be be put under the Archive page, and withe archived WIP page will no longer be updated.