House Of Quality Maker v0.0.1
A House of Quality main body is constructed from a JSON that represents all the necessary data to define a House of Quality. This is a simple proof of concept of that alone. The correlation matrix 'roof' is not included in this build. The build will be rewritten to optimize the JSON & House of Quality display to both be user-managed rather than the JSON managing the UI. The JSON is now only planned to be used for persisting the House of Quality across page loads & allowing it to be shared with a JSON data format, in addition to the already-planned image format.