House of Quality Maker v0.0.3
Majorly simplified the JSON format for constructing the House of Quality. Also added a DOM data object that holds references to the DOM elements that represent all corresponding fields in the data JSON. This will allow data on both the UI and the sava-data JSON to be handled in nearly the same way.