House of Quality Maker v0.0.5
Added a layer of separation between the elements that represent the House of Quality data and the elements that hold that data within the table. Each table cell now contains a div that holds its data. Table cells are now in a UI section of the DOM data handles, and the contents of those cells are now in a separate content section of the DOM data handles. Both of these sections are shaped like the House of Quality data, and they contain, at minimum, all of the same keys as the basic House of Quality data that's displayed to the page. This is planned to make features for data editing & content styling easier to handle in isolation from one another.