House of Quality Maker v0.0.6
Added initial export functionality, for saving the House of Quality as an image. Since the image needs to work as a standalone file before it can be delivered, the render needs to continue looking the same after it's de-coupled from the styles applied by the DOM. As is, the save-as-image functionality downloads an image where the actual body of the diagram is too small. Cloning the styles into the intermediary SVG didn't fix this issue, so investigation will continue until the image render works correctly.