House of Quality Maker v0.0.7
The HTML-to-image pipeline is being entirely reworked due to incompatibilities between browser implementations of SVGs with foreignObjects containing elements with writing-mode sideways-lr whenever those SVGs are used as the source attribute for downloadable images. For now, the page contains a testing button to view the image that's generated rather than directly downloading it. The generated image is now cross-browser compatible due to being constructed out of standard SVG shapes, instead of using the foreignObject document embed that has non-standard implementation details across browsers.